PCAクラスの定義？
